Move constants outside companion object
