+++ app/build.gradle.kts
@@ -1,11 +1,3 @@
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.ANDROIDX_APPCOMPAT
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.KOTLIN_STDLIB
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.LIFECYCLE
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.LIFECYCLE_COMPILER
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.VERSION_COMPILE_SDK
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.VERSION_MIN_SDK
-import com.github.hadilq.build.plugin.BuildPlugin.Companion.VERSION_TARGET_SDK
-
 /**
  * Copyright 2020 Hadi Lashkari Ghouchani
 
@@ -21,6 +13,14 @@ import com.github.hadilq.build.plugin.BuildPlugin.Companion.VERSION_TARGET_SDK
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import com.github.hadilq.build.plugin.ANDROIDX_APPCOMPAT
+import com.github.hadilq.build.plugin.KOTLIN_STDLIB
+import com.github.hadilq.build.plugin.LIFECYCLE
+import com.github.hadilq.build.plugin.LIFECYCLE_COMPILER
+import com.github.hadilq.build.plugin.VERSION_COMPILE_SDK
+import com.github.hadilq.build.plugin.VERSION_MIN_SDK
+import com.github.hadilq.build.plugin.VERSION_TARGET_SDK
+
 plugins {
     id("com.android.application")
     kotlin("android")
