+++ settings.gradle.kts
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+enableFeaturePreview("GRADLE_METADATA")
 pluginManagement {
     repositories {
         gradlePluginPortal()
